fix(lens): rename RemoteHostsNotice.tsx to match its RemoteHostsAlphaNotice export
git mv the file so the path carries the load-bearing "Alpha" word its primary export name does, and repoint the two import specifiers (HostSelectorStrip, MobileHostRow).

Agreed by the lowy ⇄ hickey lens debate (finding hickey-4, raised by hickey). Not pushed or merged.
